+++ database/migrations/2018_08_06_093030_create_detil_pelunasan_piutangs_table.php
@@ -18,7 +18,7 @@ class CreateDetilPelunasanPiutangsTable extends Migration
           $table->integer('bayar');
           $table->integer('discount');
           $table->string('writeoff');
-          $table->integer('piutang');
+          $table->bigInteger('piutang');
 
           $table->unsignedInteger('fj_id');
           $table->foreign('fj_id')
